H3: rework server example for clarity, remove helpers
